Fix linux compiler warnings.

+++ js_builtins.c
@@ -119,6 +119,6 @@ value_t eval_builtin(Node *a, environment_t *env) {
 	name = builtIns[idx].name;
 	fn = (firstNode *)env->table;
 
-	fprintf (stderr, "File: %s, Line: %lld, Function: %s Status: %s\n", fn->string, a->lineno, name, strstatus(v.status));
+	fprintf (stderr, "File: %s, Line: %d, Function: %s Status: %s\n", fn->string, (int)a->lineno, name, strstatus(v.status));
 	exit(1); 
 }
