+++ js_symbols.c
@@ -31,7 +31,7 @@ uint32_t insertSymbol(char *name, uint32_t len, symtab_t *symtab) {
 	symbol_t sym, *psym;
 
 	if (debug)
-		printf("insertSymbol('%.*s', %llu)\n", len, name, (uint64_t)symtab);
+		printf("insertSymbol('%.*s', %" PRIu64 ")\n", len, name, (uint64_t)symtab);
 
 	if ((psym = lookupSymbol(name, len, symtab)))
 		if (psym->depth == symtab->depth)
@@ -135,7 +135,7 @@ void hoistSymbols(uint32_t slot, hoistParams *hp) {
 	}
 	default:
 		if (debug)
-			printf("node %d unprocessed: %lld\n", slot, hp->table[slot].type);
+			printf("node %d unprocessed: %d\n", slot, (int)hp->table[slot].type);
 
 		return;
 	}
@@ -239,7 +239,7 @@ void assignSlots(uint32_t slot, Node *table, symtab_t *symtab, uint32_t depth)
 
 		if (!symbol) {
 			firstNode *fn = (firstNode *)table;
-			printf("%s: Symbol not found: %s line = %lld node = %d\n", fn->string, name->string, sym->hdr->lineno, slot);
+			printf("%s: Symbol not found: %s line = %d node = %d\n", fn->string, name->string, (int)sym->hdr->lineno, slot);
 			exit(1);
 		}
 
@@ -271,7 +271,7 @@ void assignSlots(uint32_t slot, Node *table, symtab_t *symtab, uint32_t depth)
 
 		if (idx < 0) {
 			firstNode *fn = (firstNode *)table;
-			printf("%s: Function not found: %s line = %lld node = %d\n", fn->string, name->string, sym->hdr->lineno, slot);
+			printf("%s: Function not found: %s line = %d node = %d\n", fn->string, name->string, (int)sym->hdr->lineno, slot);
 			exit(1);
 		}
 
@@ -281,7 +281,7 @@ void assignSlots(uint32_t slot, Node *table, symtab_t *symtab, uint32_t depth)
 	}
 	default:
 		if (debug)
-			printf("node type %lld skipped\n", table[slot].type);
+			printf("node type %d skipped\n", (int)table[slot].type);
 		return;
 	}
 }
